mcp/requirements.txt
# Core dependencies
mem0ai>=1.0.0
langchain-anthropic>=0.1.0
langgraph>=0.2.0

# Vector stores
chromadb>=0.4.0
lancedb>=0.3.0

# Embeddings
sentence-transformers>=2.2.0

# Utilities
python-dotenv>=1.0.0
fastmcp>=0.1.0
